Change 03.10.2021

+++ lecture01/Task2.py
@@ -1,8 +1,9 @@
 #Задание переменных
 tempC=input("Введите температуру в градусах Цельсия: ")
+
 #Программа
 try:
-    tempC=float(tempC)
+    tempC=int(tempC)
     tempF=1.8*tempC+32
     print("Температура в градусах Фаренгейта: ", tempF,"F")
 except:
